Agregar nuevos DTOs y optimizar la gestión de filtros, eliminando código obsoleto y mejorando la estructura de exportación de componentes.

+++ src/infraestructure/store/services/hotelRoomTripDetails/hotelRoomTripDetails.service.ts
@@ -5,8 +5,6 @@ import type { ApiResponse } from "../response";
 import {
   getHotelRoomTripDetailsDto,
   type GetHotelRoomTripDetailsDto,
-  hotelRoomTripDetailsDto,
-  type HotelRoomTripDetailsDto,
   type InsertManyHotelRoomTripDetailsDto,
   insertManyHotelRoomTripDetailsDto,
   updateManyHotelRoomTripDetailsByDateDto,
@@ -14,13 +12,14 @@ import {
 } from "@/domain/dtos/hotelRoomTripDetails";
 import { startShowApiError, startShowSuccess } from "@/core/utils";
 import { dateFnsAdapter } from "@/core/adapters";
-import { onSetHotelRoomTripDetails } from "../../slices/hotelRoomTripDetails.slice";
-import { onSetCurrentVersionQuotation } from "../../slices/versionQuotation.slice";
+import { hotelRoomTripDetailsCache } from "./hotelRoomTripDetails.cache";
+import { versionQuotationCache } from "../versionQuotation/versionQuotation.cache";
+import type { AppState } from "@/app/store";
 
 const PREFIX = "/hotel-room-trip-details";
 
 export const hotelRoomTripDetailsService = createApi({
-  reducerPath: "HotelRoomTripDetailsService",
+  reducerPath: "hotelRoomTripDetailsService",
   tagTypes: ["HotelRoomTripDetails", "HotelRoomTripDetail"],
   baseQuery: requestConfig(PREFIX),
   endpoints: (builder) => ({
@@ -29,22 +28,15 @@ export const hotelRoomTripDetailsService = createApi({
       GetHotelRoomTripDetailsDto
     >({
       query: (params) => {
+        const [_, errors] = getHotelRoomTripDetailsDto.create(params);
+        if (errors) throw errors;
         return {
           url: "/",
           params,
         };
       },
       providesTags: ["HotelRoomTripDetails"],
-      async onQueryStarted(params, { dispatch, queryFulfilled }) {
-        const [_, errors] = getHotelRoomTripDetailsDto.create(params);
-        if (errors) throw errors;
-        try {
-          const { data } = await queryFulfilled;
-          dispatch(onSetHotelRoomTripDetails(data.data));
-        } catch (error) {
-          throw error;
-        }
-      },
+
       transformResponse: (
         response: ApiResponse<HotelRoomTripDetailsEntity[]>
       ) => {
@@ -59,56 +51,35 @@ export const hotelRoomTripDetailsService = createApi({
         };
       },
     }),
-    createHotelRoomTripDetails: builder.mutation<
-      ApiResponse<HotelRoomTripDetailsEntity>,
-      HotelRoomTripDetailsDto
-    >({
-      query: (hotelRoomTripDetailsDto) => ({
-        url: "/",
-        method: "POST",
-        body: hotelRoomTripDetailsDto,
-      }),
-      invalidatesTags: ["HotelRoomTripDetails"],
 
-      async onQueryStarted(body, { dispatch, queryFulfilled }) {
-        const [_, errors] = hotelRoomTripDetailsDto.create(body);
-        if (errors) throw errors;
-        try {
-          const { data } = await queryFulfilled;
-          dispatch(updateHotelRoomTripDetailsCache(data.data));
-          startShowSuccess(data.message);
-        } catch (error: any) {
-          startShowApiError(error.error);
-          throw error;
-        }
-      },
-    }),
     createManyHotelRoomTripDetails: builder.mutation<
       ApiResponse<HotelRoomTripDetailsEntity[]>,
       InsertManyHotelRoomTripDetailsDto
     >({
-      query: (hotelRoomTripDetailsDto) => ({
-        url: "/many",
-        method: "POST",
-        body: hotelRoomTripDetailsDto,
-      }),
-      invalidatesTags: ["HotelRoomTripDetails"],
-
-      async onQueryStarted(body, { dispatch, queryFulfilled }) {
+      query: (body) => {
         const [_, errors] = insertManyHotelRoomTripDetailsDto.create(body);
         if (errors) throw errors;
+        return {
+          url: "/many",
+          method: "POST",
+          body,
+        };
+      },
+      async onQueryStarted(_, { queryFulfilled, dispatch, getState }) {
         try {
           const { data } = await queryFulfilled;
-          data.data.forEach((hotelRoomTripDetails) => {
-            dispatch(updateHotelRoomTripDetailsCache(hotelRoomTripDetails));
-            if (hotelRoomTripDetails.tripDetails) {
-              dispatch(
-                onSetCurrentVersionQuotation(
-                  hotelRoomTripDetails.tripDetails?.versionQuotation ?? null
-                )
-              );
-            }
-          });
+          hotelRoomTripDetailsCache.createManyHotelRoomTripDetails(
+            data,
+            dispatch,
+            getState
+          );
+
+          versionQuotationCache.addMultipleHotelRoomTripDetails(
+            data.data,
+            dispatch,
+            getState as () => AppState
+          );
+
           startShowSuccess(data.message);
         } catch (error: any) {
           if (error.error) startShowApiError(error.error);
@@ -121,24 +92,26 @@ export const hotelRoomTripDetailsService = createApi({
       ApiResponse<HotelRoomTripDetailsEntity[]>,
       UpdateManyHotelRoomTripDetailsByDateDto
     >({
-      query: (params) => ({
-        url: "/many/date",
-        method: "PUT",
-        body: params,
-      }),
-      invalidatesTags: ["HotelRoomTripDetails"],
-      async onQueryStarted(body, { dispatch, queryFulfilled }) {
+      query: (body) => {
+        const [_, errors] =
+          updateManyHotelRoomTripDetailsByDateDto.create(body);
+        if (errors) throw errors;
+        return {
+          url: "/many/date",
+          method: "PUT",
+          body,
+        };
+      },
+      async onQueryStarted(_, { queryFulfilled, dispatch, getState }) {
         try {
-          const [_, errors] =
-            updateManyHotelRoomTripDetailsByDateDto.create(body);
-          if (errors) throw errors;
           const { data } = await queryFulfilled;
-          data.data.forEach((hotelRoomTripDetails) => {
-            dispatch(updateHotelRoomTripDetailsCache(hotelRoomTripDetails));
-          });
-          // startShowSuccess(data.message);
+          versionQuotationCache.updateManyHotelRoomTripDetailsByDate(
+            data.data,
+            dispatch,
+            getState as () => AppState
+          );
         } catch (error: any) {
-          startShowApiError(error.error);
+          if (error.error) startShowApiError(error.error);
           throw error;
         }
       },
@@ -146,25 +119,20 @@ export const hotelRoomTripDetailsService = createApi({
 
     deleteHotelRoomTripDetails: builder.mutation<
       ApiResponse<HotelRoomTripDetailsEntity>,
-      number
+      HotelRoomTripDetailsEntity["id"]
     >({
       query: (id) => ({
         url: `/${id}`,
         method: "DELETE",
       }),
-      invalidatesTags: ["HotelRoomTripDetails"],
-      async onQueryStarted(_, { dispatch, queryFulfilled }) {
+      async onQueryStarted(_, { dispatch, queryFulfilled, getState }) {
         try {
           const { data } = await queryFulfilled;
-          dispatch(deleteHotelRoomTripDetailsCache(data.data));
-          // startShowSuccess(data.message);
-          if (data.data.tripDetails) {
-            dispatch(
-              onSetCurrentVersionQuotation(
-                data.data.tripDetails?.versionQuotation ?? null
-              )
-            );
-          }
+          versionQuotationCache.deleteHotelRoomTripDetails(
+            data.data,
+            dispatch,
+            getState as () => AppState
+          );
         } catch (error: any) {
           startShowApiError(error.error);
           throw error;
@@ -174,35 +142,33 @@ export const hotelRoomTripDetailsService = createApi({
 
     deleteManyHotelRoomTripDetails: builder.mutation<
       ApiResponse<HotelRoomTripDetailsEntity[]>,
-      number[]
+      HotelRoomTripDetailsEntity["id"][]
     >({
-      query: (ids) => ({
-        url: "/many/date",
-        method: "DELETE",
-        body: {
-          ids,
-        },
-        headers: {
-          "Content-Type": "application/json",
-        },
-      }),
-      invalidatesTags: ["HotelRoomTripDetails"],
-      async onQueryStarted(_, { dispatch, queryFulfilled }) {
+      query: (ids) => {
+        if (!ids.length) {
+          throw "No ids provided";
+        }
+        return {
+          url: "/many/date",
+          method: "DELETE",
+          body: {
+            ids,
+          },
+          headers: {
+            "Content-Type": "application/json",
+          },
+        };
+      },
+      async onQueryStarted(_, { dispatch, queryFulfilled, getState }) {
         try {
           const { data } = await queryFulfilled;
-          data.data.forEach((hotelRoomTripDetails) => {
-            dispatch(deleteHotelRoomTripDetailsCache(hotelRoomTripDetails));
-            if (hotelRoomTripDetails.tripDetails) {
-              dispatch(
-                onSetCurrentVersionQuotation(
-                  hotelRoomTripDetails.tripDetails?.versionQuotation ?? null
-                )
-              );
-            }
-          });
-          // startShowSuccess(data.message);
+          versionQuotationCache.deleteManyHotelRoomTripDetails(
+            data.data,
+            dispatch,
+            getState as () => AppState
+          );
         } catch (error: any) {
-          // startShowApiError(error.error);
+          startShowApiError(error.error);
           throw error;
         }
       },
@@ -212,45 +178,8 @@ export const hotelRoomTripDetailsService = createApi({
 
 export const {
   useGetAllHotelRoomTripDetailsQuery,
-  useCreateHotelRoomTripDetailsMutation,
   useCreateManyHotelRoomTripDetailsMutation,
   useUpdateManyHotelRoomTripDetailsByDateMutation,
   useDeleteHotelRoomTripDetailsMutation,
   useDeleteManyHotelRoomTripDetailsMutation,
 } = hotelRoomTripDetailsService;
-
-//* Cache update
-const updateHotelRoomTripDetailsCache = (
-  hotelRoomTripDetails: HotelRoomTripDetailsEntity
-) => {
-  return hotelRoomTripDetailsService.util.updateQueryData(
-    "getAllHotelRoomTripDetails",
-    {
-      tripDetailsId: hotelRoomTripDetails.id,
-    },
-    (draft) => {
-      if (draft?.data) {
-        draft.data.push(hotelRoomTripDetails);
-      }
-    }
-  );
-};
-
-const deleteHotelRoomTripDetailsCache = (
-  hotelRoomTripDetails: HotelRoomTripDetailsEntity
-) => {
-  return hotelRoomTripDetailsService.util.updateQueryData(
-    "getAllHotelRoomTripDetails",
-    {
-      tripDetailsId: hotelRoomTripDetails.id,
-    },
-    (draft) => {
-      if (draft?.data) {
-        const index = draft.data.findIndex(
-          (c) => c.id === hotelRoomTripDetails.id
-        );
-        if (index !== -1) draft.data.splice(index, 1);
-      }
-    }
-  );
-};
